refactoring into utils file
